+++ src/hooks/usePreventScroll.tsx
@@ -3,59 +3,48 @@ import { UsePreventScrollProps } from '../sheetType';
 
 const usePreventScroll = ({
   scrollRef,
+  footerRef,
   bottomScrollLock,
   position,
 }: UsePreventScrollProps) => {
-  const isContentTouchedRef = useRef(false);
   const hasScrolledRef = useRef(false);
   const initialTouchYCoordRef = useRef(0);
   const lastScrollTopRef = useRef<number>(0);
-
   // 요소가 스크롤을 가지고 있는지 판단하는 함수
   const doesElementHaveScroll = (element: HTMLElement) => {
     return element.scrollHeight > element.clientHeight;
   };
-
   // TODO [ ] - 불필요한 변수 제거하기
   // 위로 슬라이드 시 이벤트를 방지할지 결정하는 함수
   const shouldPreventOnSlideUp = useCallback(
     (hasScroll: boolean, isTopPosition: boolean) => {
       // 스크롤이 없고, 현재 위치가 'top'인지 확인
       const isNoScrollAndTop = !hasScroll && isTopPosition;
-      // content가 터치되었고, 스크롤이 없는지 확인
-      const isTouchedNoScrollContent =
-        isContentTouchedRef.current && !hasScroll;
       // 바텀일 때 스크롤이 잠금이 true인지, 현재 위치가 'bottom'인지 확인
       const isBottomLocked = bottomScrollLock && position === 'bottom';
-      return isNoScrollAndTop || isTouchedNoScrollContent || isBottomLocked;
+      return isNoScrollAndTop || isBottomLocked;
     },
     [bottomScrollLock, position]
   );
-
   useEffect(() => {
     if (!scrollRef.current) return;
-
     const handleTouchStart = (e: TouchEvent) => {
       if (!scrollRef.current) return;
       initialTouchYCoordRef.current = e.touches[0]?.clientY ?? 0; // 터치 시작점의 Y 좌표 저장
       hasScrolledRef.current = doesElementHaveScroll(scrollRef.current); // 세로 스크롤 여부를 판단
-      isContentTouchedRef.current = true; // 컨텐츠 터치 여부
     };
-
-    const handleTouchEnd = () => {
+    const handleTouchEnd = (e: TouchEvent) => {
       if (!scrollRef.current) return;
       hasScrolledRef.current = doesElementHaveScroll(scrollRef.current); // 세로 스크롤 여부를 재확인
-      isContentTouchedRef.current = false; // 컨텐츠 터치 여부
     };
-
     const handleTouchMove = (e: TouchEvent) => {
       if (!scrollRef.current || !e.touches[0]) return;
-
       const touchEndY = e.touches[0].clientY;
       const isSlideDown = touchEndY > initialTouchYCoordRef.current;
       const shouldPreventDown =
-        !hasScrolledRef.current || scrollRef.current.scrollTop <= 0;
-
+        !hasScrolledRef.current ||
+        scrollRef.current.scrollTop <= 0 ||
+        (bottomScrollLock && position === 'bottom');
       // 위로 움직였을 때, 스크롤이 없거나 현재 위치가 'top'이거나 바텀 스크롤 잠금이 있는 경우 위로 스크롤 막기
       const shouldPreventUP = shouldPreventOnSlideUp(
         hasScrolledRef.current,
@@ -68,16 +57,14 @@ const usePreventScroll = ({
         e.preventDefault();
       }
     };
-
     let prevValue = 0;
     // Safari에서의 오버스크롤을 방지하는 함수
-    const preventSafariOverscrollOnStart = () => {
+    const preventSafariOverscrollOnStart = (e: TouchEvent) => {
       if (!scrollRef.current) return;
       if (scrollRef.current.scrollTop <= 0) {
         prevValue = scrollRef.current.scrollTop;
       }
     };
-
     // Safari에서의 오버스크롤을 방지하는 함수
     const preventSafariOverscrollOnMove = (e: TouchEvent) => {
       if (!scrollRef.current) return;
@@ -95,6 +82,21 @@ const usePreventScroll = ({
       }
     };
 
+    const bottomLockScrollEvent = (e: Event) => {
+      if (bottomScrollLock && position === 'bottom') {
+        e.preventDefault();
+      }
+    };
+    //footerElement 이벤트 막기
+    const footerPrevent = (e: TouchEvent) => {
+      e.preventDefault();
+    };
+    if (footerRef.current) {
+      footerRef.current.addEventListener('touchmove', footerPrevent, {
+        passive: false,
+      });
+    }
+
     scrollRef.current.addEventListener('touchstart', handleTouchStart, {
       passive: true,
     });
@@ -109,7 +111,6 @@ const usePreventScroll = ({
         passive: false,
       }
     );
-
     scrollRef.current.addEventListener(
       'touchstart',
       preventSafariOverscrollOnStart,
@@ -119,6 +120,7 @@ const usePreventScroll = ({
     );
 
     scrollRef.current.addEventListener('scroll', recordScrollTop);
+    scrollRef.current.addEventListener('wheel', bottomLockScrollEvent);
 
     return () => {
       if (!scrollRef.current) return;
@@ -134,15 +136,22 @@ const usePreventScroll = ({
         preventSafariOverscrollOnStart
       );
       scrollRef.current.removeEventListener('scroll', recordScrollTop);
+      scrollRef.current.removeEventListener('wheel', bottomLockScrollEvent);
+      if (footerRef.current) {
+        footerRef.current.removeEventListener('touchmove', footerPrevent);
+      }
     };
-  }, [scrollRef, bottomScrollLock, position, shouldPreventOnSlideUp]);
-
+  }, [
+    scrollRef,
+    bottomScrollLock,
+    position,
+    shouldPreventOnSlideUp,
+    footerRef,
+  ]);
   return {
-    isContentTouchedRef,
     hasScrolledRef,
     initialTouchYCoordRef,
     lastScrollTopRef,
   };
 };
-
 export { usePreventScroll };